+++ src/python/zensols/zotsite/__init__.py
@@ -1,5 +1,4 @@
 from .domain import *
-from .dbreader import *
 from .itemmap import *
 from .navvisitor import *
 from .fscpvisitor import *